% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PakPC2023PakDist.R
\docType{data}
\name{PakPC2023PakDist}
\alias{PakPC2023PakDist}
\title{District data from Pakistan Population Census 2023}
\format{
A \code{data.table} and  \code{data.frame}.
\describe{
       \item{\code{Region}}{Region}
       \item{\code{Division}}{Division}
       \item{\code{District}}{District}
       \item{\code{Households}}{No. of Households}
       \item{\code{Pop2023}}{Total Population in 2023}
       \item{\code{AHS}}{Average Household Size: Pop2023/Households}
       \item{\code{Pop2017}}{Total Population in 2017}
       \item{\code{GR}}{Average Annual Population Growth Rate: ((Pop2023/Pop2017)^(1/6)-1)*100}
       }
}
\usage{
data(PakPC2023PakDist)
}
\description{
\code{PakPC2023PakDist} Pakistan data from Pakistan Population Census 2023.
}
\examples{
data(PakPC2023PakDist)
}
\references{
\enumerate{
\item Pakistan Population Census 2023 (\url{https://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{PakPC2023Pak}}
 , \code{\link{PakPC2023PakDiv}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\keyword{datasets}
