% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_01.R
\docType{data}
\name{TABLE_01}
\alias{TABLE_01}
\title{Area, population by sex, sex ratio, population density, urban proportion, household size and annual growth rate}
\format{
A \code{data.table} and  \code{data.frame}.
\describe{
 \item{\code{NAME_ADMIN_UNIT}}{NAME OF ADMINISTRATIVE UNIT}
 \item{\code{ADMIN_UNIT}}{ADMIN UNIT}
 \item{\code{PART_OF_DISTRICT}}{PART OF DISTRICT}
 \item{\code{REGION}}{REGION}
 \item{\code{AREA_SQKM}}{AREA IN SQ KM}
 \item{\code{ALL_SEXES}}{ALL SEXES}
 \item{\code{MALE}}{MALE}
 \item{\code{FEMALE}}{FEMALE}
 \item{\code{TGEND}}{TRANSGENDER}
 \item{\code{SEX_RATIO}}{SEX RATIO}
 \item{\code{POP_DEN_SQKM}}{DENSITY PER SQUARE KM}
 \item{\code{URBAN_PROP}}{URBAN PROPORTION}
 \item{\code{AVG_HH_SIZE}}{AVERAGE HOUSEHOLD SIZE}
 \item{\code{POP_2017}}{POPULATION 2017}
 \item{\code{AVG_ANNUAL_GR_RATE_17_23}}{2017-2023 AVERAGE ANNUAL GROWTH RATE}
 \item{\code{PART_OF_DIVISION}}{PART OF DIVISION}
 \item{\code{PART_OF_PROVINCE}}{PART OF PROVINCE}
       }
}
\usage{
data(TABLE_01)
}
\description{
\code{TABLE_01} Area, population by sex, sex ratio, population density, urban proportion, household size and annual growth rate
}
\examples{
data(TABLE_01)
}
\references{
\enumerate{
\item Pakistan Population Census 2023 (\url{https://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{PakPC2023PakDiv}}
 , \code{\link{PakPC2023PakDist}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\keyword{datasets}
