% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_02.R
\docType{data}
\name{TABLE_02}
\alias{TABLE_02}
\title{Urban localities by population size and their population by sex, annual growth rate and household size}
\format{
A \code{data.table} and  \code{data.frame}.
\describe{
  \item{\code{PROVINCE}}{PROVINCE}
  \item{\code{DIVISION}}{DIVISION}
  \item{\code{DISTRICT}}{DISTRICT}
  \item{\code{TEHSIL}}{TEHSIL}
  \item{\code{ADMIN_UNIT}}{ADMIN UNIT}
  \item{\code{REGION}}{REGION}
  \item{\code{AREA_SQKM}}{AREA IN SQ KM}
  \item{\code{ALL_SEXES}}{ALL SEXES}
  \item{\code{MALE}}{MALE}
  \item{\code{FEMALE}}{FEMALE}
  \item{\code{TGEND}}{TRANSGENDER}
  \item{\code{SEX_RATIO}}{SEX RATIO}
  \item{\code{POP_DEN_SQKM}}{DENSITY PER SQUARE KM}
  \item{\code{URBAN_PROP}}{URBAN PROPORTION}
  \item{\code{AVG_HH_SIZE}}{AVERAGE HOUSEHOLD SIZE}
  \item{\code{POP_2017}}{POPULATION 2017}
  \item{\code{AVG_ANNUAL_GR_RATE_17_23}}{2017-2023 AVERAGE ANNUAL GROWTH RATE}
       }
}
\usage{
data(TABLE_02)
}
\description{
\code{TABLE_02} Urban localities by population size and their population by sex, annual growth rate and household size
}
\examples{
data(TABLE_02)
}
\references{
\enumerate{
\item Pakistan Population Census 2023 (\url{https://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{PakPC2023PakDiv}}
 , \code{\link{PakPC2023PakDist}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\keyword{datasets}
