% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_03.R
\docType{data}
\name{TABLE_03}
\alias{TABLE_03}
\title{Number of rural localities by population size and their population by sex}
\format{
A \code{data.table} and  \code{data.frame}.
\describe{
  \item{\code{PROVINCE}}{PROVINCE}
  \item{\code{DIVISION}}{DIVISION}
  \item{\code{DISTRICT}}{DISTRICT}
  \item{\code{TEHSIL}}{TEHSIL}
  \item{\code{ADMIN_UNIT}}{ADMIN UNIT}
  \item{\code{REGION}}{REGION}
  \item{\code{RURAL_LOCALITIES_BY_POP_SIZE}}{RURAL LOCALITIES BY POP SIZE}
  \item{\code{NUMBER_OF_RURAL_LOCALITIES}}{NUMBER OF RURAL LOCALITIES}
  \item{\code{ALL_SEXES}}{ALL SEXES}
  \item{\code{MALE}}{MALE}
  \item{\code{FEMALE}}{FEMALE}
  \item{\code{TGEND}}{TRANSGENDER}
  \item{\code{NOTE}}{NOTE}
       }
}
\usage{
data(TABLE_03)
}
\description{
\code{TABLE_03} Number of rural localities by population size and their population by sex
}
\examples{
data(TABLE_03)
}
\references{
\enumerate{
\item Pakistan Population Census 2023 (\url{https://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{PakPC2023PakDiv}}
 , \code{\link{PakPC2023PakDist}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\keyword{datasets}
