% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_06.R
\docType{data}
\name{TABLE_06}
\alias{TABLE_06}
\title{Population (15 years and above) by age group, sex, marital status and rural/urban}
\format{
A \code{data.table} and  \code{data.frame}.
\describe{
  \item{\code{PROVINCE}}{PROVINCE}
  \item{\code{DISTRICT}}{DISTRICT}
  \item{\code{AGE_GROUP}}{AGE GROUP}
  \item{\code{MARITAL_STATUS}}{MARITAL STATUS}
  \item{\code{ALL_SEXES_OVERALL}}{ALL SEXES OVERALL}
  \item{\code{FEMALE_OVERALL}}{FEMALE OVERALL}
  \item{\code{MALE_OVERALL}}{MALE OVERALL}
  \item{\code{TGEND_OVERALL}}{TRANSGENDER OVERALL}
  \item{\code{ALL_SEXES_RURAL}}{ALL SEXES RURAL}
  \item{\code{FEMALE_RURAL}}{FEMALE RURAL}
  \item{\code{MALE_RURAL}}{MALE RURAL}
  \item{\code{TGEND_RURAL}}{TRANSGENDER RURAL}
  \item{\code{ALL_SEXES_URBAN}}{ALL_SEXES_URBAN}
  \item{\code{FEMALE_URBAN}}{FEMALE_URBAN}
  \item{\code{MALE_URBAN}}{MALE_URBAN}
  \item{\code{TGEND_URBAN}}{TGEND_URBAN}
       }
}
\usage{
data(TABLE_06)
}
\description{
\code{TABLE_06} Population (15 years and above) by age group, sex, marital status and rural/urban
}
\examples{
data(TABLE_06)
}
\references{
\enumerate{
\item Pakistan Population Census 2023 (\url{https://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{PakPC2023PakDiv}}
 , \code{\link{PakPC2023PakDist}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\keyword{datasets}
