% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_11.R
\docType{data}
\name{TABLE_11}
\alias{TABLE_11}
\title{Population by mother tongue, sex and rural/urban}
\format{
A \code{data.table} and  \code{data.frame}.
\describe{
  \item{\code{PROVINCE}}{PROVINCE}
  \item{\code{DIVISION}}{DIVISION}
  \item{\code{DISTRICT}}{DISTRICT}
  \item{\code{TEHSIL}}{TEHSIL}
  \item{\code{ADMIN_UNIT}}{ADMIN UNIT}
  \item{\code{LANGUAGE}}{LANGUAGE}
  \item{\code{ALL_SEXES_OVERALL}}{ALL SEXES OVERALL}
  \item{\code{FEMALE_OVERALL}}{FEMALE OVERALL}
  \item{\code{MALE_OVERALL}}{MALE OVERALL}
  \item{\code{TGEND_OVERALL}}{TRANSGENDER OVERALL}
  \item{\code{ALL_SEXES_RURAL}}{ALL SEXES RURAL}
  \item{\code{FEMALE_RURAL}}{FEMALE RURAL}
  \item{\code{MALE_RURAL}}{MALE RURAL}
  \item{\code{TGEND_RURAL}}{TRANSGENDER RURAL}
  \item{\code{ALL_SEXES_URBAN}}{ALL SEXES URBAN}
  \item{\code{FEMALE_URBAN}}{FEMALE URBAN}
  \item{\code{MALE_URBAN}}{MALE URBAN}
  \item{\code{TGEND_URBAN}}{TRANSGENDER URBAN}
       }
}
\usage{
data(TABLE_11)
}
\description{
\code{TABLE_11} Population by mother tongue, sex and rural/urban
}
\examples{
data(TABLE_11)
}
\references{
\enumerate{
\item Pakistan Population Census 2023 (\url{https://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{PakPC2023PakDiv}}
 , \code{\link{PakPC2023PakDist}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\keyword{datasets}
