% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_16.R
\docType{data}
\name{TABLE_16}
\alias{TABLE_16}
\title{Disability and functional limitation by region and gender}
\format{
A \code{data.table} and  \code{data.frame}.
\describe{
  \item{\code{PROVINCE}}{PROVINCE}
  \item{\code{DIVISION}}{DIVISION}
  \item{\code{DISTRICT}}{DISTRICT}
  \item{\code{TEHSIL}}{TEHSIL}
  \item{\code{ADMIN_UNIT}}{ADMIN UNIT}
  \item{\code{DISAB_FUNC_LIM}}{DISAB FUNC LIM}
  \item{\code{ALL_SEXES_OVERALL}}{ALL SEXES OVERALL}
  \item{\code{MALE_OVERALL}}{MALE OVERALL}
  \item{\code{FEMALE_OVERALL}}{FEMALE OVERALL}
  \item{\code{ALL_SEXES_RURAL}}{ALL SEXES RURAL}
  \item{\code{MALE_RURAL}}{MALE RURAL}
  \item{\code{FEMALE_RURAL}}{FEMALE RURAL}
  \item{\code{ALL_SEXES_URBAN}}{ALL SEXES URBAN}
  \item{\code{MALE_URBAN}}{MALE URBAN}
  \item{\code{FEMALE_URBAN}}{FEMALE URBAN}
       }
}
\usage{
data(TABLE_16)
}
\description{
\code{TABLE_16} Disability and functional limitation by region and gender
}
\examples{
data(TABLE_16)
}
\references{
\enumerate{
\item Pakistan Population Census 2023 (\url{https://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{PakPC2023PakDiv}}
 , \code{\link{PakPC2023PakDist}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\keyword{datasets}
