% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_17.R
\docType{data}
\name{TABLE_17}
\alias{TABLE_17}
\title{Disability and functional limitation for special age group by rural/urban}
\format{
A \code{data.table} and  \code{data.frame}.
\describe{
  \item{\code{PROVINCE}}{PROVINCE}
  \item{\code{DIVISION}}{DIVISION}
  \item{\code{DISTRICT}}{DISTRICT}
  \item{\code{TEHSIL}}{TEHSIL}
  \item{\code{ADMIN_UNIT}}{ADMIN UNIT}
  \item{\code{DISAB_FUNC_LIM}}{DISAB FUNC LIM}
  \item{\code{AGE_GROUP}}{AGE GROUP}
  \item{\code{OVERALL}}{OVERALL}
  \item{\code{RURAL}}{RURAL}
  \item{\code{URBAN}}{URBAN}
       }
}
\usage{
data(TABLE_17)
}
\description{
\code{TABLE_17} Disability and functional limitation for special age group by rural/urban
}
\examples{
data(TABLE_17)
}
\references{
\enumerate{
\item Pakistan Population Census 2023 (\url{https://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{PakPC2023PakDiv}}
 , \code{\link{PakPC2023PakDist}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\keyword{datasets}
