% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_20.R
\docType{data}
\name{TABLE_20}
\alias{TABLE_20}
\title{Type of housing unit by region}
\format{
A \code{data.table} and  \code{data.frame}.
\describe{
  \item{\code{PROVINCE}}{PROVINCE}
  \item{\code{DIVISION}}{DIVISION}
  \item{\code{DISTRICT}}{DISTRICT}
  \item{\code{TEHSIL}}{TEHSIL}
  \item{\code{ADMIN_UNIT}}{ADMIN UNIT}
  \item{\code{HH_TYPE}}{HH TYPE}
  \item{\code{OVERALL}}{OVERALL}
  \item{\code{RURAL}}{RURAL}
  \item{\code{URBAN}}{URBAN}
       }
}
\usage{
data(TABLE_21)
}
\description{
\code{TABLE_21} Type of housing unit by region
}
\examples{
data(TABLE_21)
}
\references{
\enumerate{
\item Pakistan Population Census 2023 (\url{https://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{PakPC2023PakDiv}}
 , \code{\link{PakPC2023PakDist}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\keyword{datasets}
