% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_21.R
\docType{data}
\name{TABLE_21}
\alias{TABLE_21}
\title{Types of households by population and sex, rural/urban}
\format{
A \code{data.table} and  \code{data.frame}.
\describe{
  \item{\code{PROVINCE}}{PROVINCE}
  \item{\code{DIVISION}}{DIVISION}
  \item{\code{DISTRICT}}{DISTRICT}
  \item{\code{TEHSIL}}{TEHSIL}
  \item{\code{ADMIN_UNIT}}{ADMIN UNIT}
  \item{\code{REGION}}{REGION}
  \item{\code{TYPE_OF_HOUSING_UNITS}}{TYPE OF HOUSING UNITS}
  \item{\code{ALL_SEXES}}{ALL SEXES}
  \item{\code{MALE}}{MALE}
  \item{\code{FEMALE}}{FEMALE}
  \item{\code{TGEND}}{TGEND}
  \item{\code{HOUSING_UNITS_TOTAL}}{HOUSING UNITS TOTAL}
  \item{\code{HOUSING_UNITS_PCT}}{HOUSING UNITS PCT}
       }
}
\usage{
data(TABLE_21)
}
\description{
\code{TABLE_21} Types of households by population and sex, rural/urban
}
\examples{
data(TABLE_21)
}
\references{
\enumerate{
\item Pakistan Population Census 2023 (\url{https://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{PakPC2023PakDiv}}
 , \code{\link{PakPC2023PakDist}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\keyword{datasets}
