% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_22.R
\docType{data}
\name{TABLE_22}
\alias{TABLE_22}
\title{Housing characteristics, facilities used for fuel, lighting and kitchen by region}
\format{
A \code{data.table} and  \code{data.frame}.
\describe{
  \item{\code{PROVINCE}}{PROVINCE}
  \item{\code{DIVISION}}{DIVISION}
  \item{\code{DISTRICT}}{DISTRICT}
  \item{\code{TEHSIL}}{TEHSIL}
  \item{\code{ADMIN_UNIT}}{ADMIN UNIT}
  \item{\code{REGION}}{REGION}
  \item{\code{HOUSEHOLDS}}{HOUSEHOLDS}
  \item{\code{LIGHT_ELECT}}{LIGHT ELECT}
  \item{\code{LIGHT_SOLAR}}{LIGHT SOLAR}
  \item{\code{LIGHT_OTHERS}}{LIGHT OTHERS}
  \item{\code{FUEL_GAS}}{FUEL GAS}
  \item{\code{FUEL_LPGCNG}}{FUEL LPGCNG}
  \item{\code{FUEL_FIREWOOD}}{FUEL FIREWOOD}
  \item{\code{FUEL_OTHERS}}{FUEL OTHERS}
  \item{\code{SEP_KITCHEN}}{SEPARATE KITCHEN}
  \item{\code{NO_KITCHEN}}{NO KITCHEN}
       }
}
\usage{
data(TABLE_22)
}
\description{
\code{TABLE_22} Housing characteristics, facilities used for fuel, lighting and kitchen by region
}
\examples{
data(TABLE_22)
}
\references{
\enumerate{
\item Pakistan Population Census 2023 (\url{https://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{PakPC2023PakDiv}}
 , \code{\link{PakPC2023PakDist}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\keyword{datasets}
