% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_23.R
\docType{data}
\name{TABLE_23}
\alias{TABLE_23}
\title{Housing facilities by sources of drinking water by region}
\format{
A \code{data.table} and  \code{data.frame}.
\describe{
  \item{\code{PROVINCE}}{PROVINCE}
  \item{\code{DIVISION}}{DIVISION}
  \item{\code{DISTRICT}}{DISTRICT}
  \item{\code{TEHSIL}}{TEHSIL}
  \item{\code{ADMIN_UNIT}}{ADMIN UNIT}
  \item{\code{REGION}}{REGION}
  \item{\code{HOUSEHOLDS}}{HOUSEHOLDS}
  \item{\code{DRINK_WTR_IMPROVE}}{DRINK WATER IMPROVE}
  \item{\code{DRINK_WTR_INSIDE}}{DRINK WATER INSIDE}
  \item{\code{DRINK_WTR_OUTSIDE}}{DRINK WATER OUTSIDE}
  \item{\code{DRINK_WTR_TAP}}{DRINK WATER TAP}
  \item{\code{DRINK_WTR_MOTOR}}{DRINK WATER MOTOR}
  \item{\code{DRINK_WTR_WELL}}{DRINK WATER WELL}
  \item{\code{DRINK_WTR_FILTER}}{DRINK WATER FILTER}
  \item{\code{DRINK_WTR_BOTTLE}}{DRINK WATER BOTTLE}
  \item{\code{DRINK_WTR_OTHER}}{DRINK WATER OTHER}
       }
}
\usage{
data(TABLE_23)
}
\description{
\code{TABLE_23} Housing facilities by sources of drinking water by region
}
\examples{
data(TABLE_23)
}
\references{
\enumerate{
\item Pakistan Population Census 2023 (\url{https://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{PakPC2023PakDiv}}
 , \code{\link{PakPC2023PakDist}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\keyword{datasets}
