% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_24.R
\docType{data}
\name{TABLE_24}
\alias{TABLE_24}
\title{Housing characteristics by housing structure, type of toilet and washroom by region}
\format{
A \code{data.table} and  \code{data.frame}.
\describe{
  \item{\code{PROVINCE}}{PROVINCE}
  \item{\code{DIVISION}}{DIVISION}
  \item{\code{DISTRICT}}{DISTRICT}
  \item{\code{TEHSIL}}{TEHSIL}
  \item{\code{ADMIN_UNIT}}{ADMIN UNIT}
  \item{\code{REGION}}{REGION}
  \item{\code{HOUSEHOLDS}}{HOUSEHOLDS}
  \item{\code{TOILET_SEPARATE}}{TOILET SEPARATE}
  \item{\code{TOILET_FLUSH}}{TOILET FLUSH}
  \item{\code{TOILET_NON_FLUSH}}{TOILET NON FLUSH}
  \item{\code{TOILET_NONE}}{TOILET NONE}
  \item{\code{WASHROOM_SEPARATE}}{WASHROOM SEPARATE}
  \item{\code{WASHROOM_NONE}}{WASHROOM NONE}
       }
}
\usage{
data(TABLE_24)
}
\description{
\code{TABLE_24} Housing characteristics by housing structure, type of toilet and washroom by region
}
\examples{
data(TABLE_24)
}
\references{
\enumerate{
\item Pakistan Population Census 2023 (\url{https://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{PakPC2023PakDiv}}
 , \code{\link{PakPC2023PakDist}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\keyword{datasets}
