% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_25.R
\docType{data}
\name{TABLE_25}
\alias{TABLE_25}
\title{Housing characteristics by residential status, ownership and number of rooms by region}
\format{
A \code{data.table} and  \code{data.frame}.
\describe{
  \item{\code{PROVINCE}}{PROVINCE}
  \item{\code{DIVISION}}{DIVISION}
  \item{\code{DISTRICT}}{DISTRICT}
  \item{\code{TEHSIL}}{TEHSIL}
  \item{\code{ADMIN_UNIT}}{ADMIN UNIT}
  \item{\code{REGION}}{REGION}
  \item{\code{HOUSEHOLDS}}{HOUSEHOLDS}
  \item{\code{RES_ST_OWNED}}{RESIDENTAL STATUS OWNED}
  \item{\code{RES_ST_RENTED}}{RESIDENTAL STATUS RENTED}
  \item{\code{RES_ST_RENT_FREE}}{RESIDENTAL STATUS RENT FREE}
  \item{\code{RES_ST_GOVT}}{RESIDENTAL STATUS GOVT}
  \item{\code{RES_ST_NON_GOVT}}{RESIDENTAL STATUS NON GOVT}
  \item{\code{RES_ST_OTHER}}{RESIDENTAL STATUS OTHER}
  \item{\code{OWNER_GENDER_MALE}}{OWNER GENDER MALE}
  \item{\code{OWNER_GENDER_FEMALE}}{OWNER GENDER FEMALE}
  \item{\code{NO_ROOMS_ONE}}{ONE ROOM}
  \item{\code{NO_ROOMS_TWO}}{TWO ROOMS}
  \item{\code{NO_ROOMS_THREE}}{THREE ROOMS}
  \item{\code{NO_ROOMS_GRTREQ_FOUR}}{FOUR PLUS ROOMS}
       }
}
\usage{
data(TABLE_25)
}
\description{
\code{TABLE_25} Housing characteristics by residential status, ownership and number of rooms by region
}
\examples{
data(TABLE_25)
}
\references{
\enumerate{
\item Pakistan Population Census 2023 (\url{https://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{PakPC2023PakDiv}}
 , \code{\link{PakPC2023PakDist}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\keyword{datasets}
