% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLE_26.R
\docType{data}
\name{TABLE_26}
\alias{TABLE_26}
\title{Total number of structure by types and rural/urban}
\format{
A \code{data.table} and  \code{data.frame}.
\describe{
  \item{\code{PROVINCE}}{PROVINCE}
  \item{\code{DIVISION}}{DIVISION}
  \item{\code{DISTRICT}}{DISTRICT}
  \item{\code{TEHSIL}}{TEHSIL}
  \item{\code{ADMIN_UNIT}}{ADMIN UNIT}
  \item{\code{REGION}}{REGION}
  \item{\code{ALL_STRUCTURES}}{ALL STRUCTURES}
  \item{\code{HIGH_RISE_RES}}{HIGH RISE RESIDENTIAL}
  \item{\code{HIGH_RISE_ECON}}{HIGH RISE ECONOMIC}
  \item{\code{HIGH_RISE_RES_ECON}}{HIGH RISE RESIDENTIAL & ECONOMIC}
  \item{\code{NORMAL_RES}}{NORMAL RESIDENTIAL}
  \item{\code{NORMAL_ECON}}{NORMAL ECONOMIC}
  \item{\code{NORMAL_RES_ECON}}{NORMAL RESIDENTIAL & ECONOMIC}
  \item{\code{OTHER_JUGHI}}{OTHER JUGHI}
  \item{\code{OTHER_UNDER_CONSTR}}{OTHER UNDER CONSTRUCTION}
       }
}
\usage{
data(TABLE_26)
}
\description{
\code{TABLE_26} Total number of structure by types and rural/urban
}
\examples{
data(TABLE_26)
}
\references{
\enumerate{
\item Pakistan Population Census 2023 (\url{https://www.pbs.gov.pk/}).
 }
}
\seealso{
\code{\link{PakPC2023PakDiv}}
 , \code{\link{PakPC2023PakDist}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\keyword{datasets}
